chore: move multipart upload to native handlers
