+++ package/expo-package/src/handlers/index.ts
@@ -1,1 +1,2 @@
 export * from './compressImage';
+export * from './multipartUpload';
