+++ package/expo-package/src/handlers/multipartUpload.ts
@@ -1,8 +1,7 @@
 import { createNativeMultipartUpload } from 'stream-chat-react-native-core';
 
-import { getLocalAssetUri } from './getLocalAssetUri';
-
 import { uploadMultipart } from '../native/multipartUploader';
+import { getLocalAssetUri } from '../optionalDependencies/getLocalAssetUri';
 
 export const multipartUpload = createNativeMultipartUpload({
   getLocalAssetUri,
